refs/a2a-latency-extension/samples/python/requirements.txt
fastapi>=0.115.14
pydantic>=2.11.7
uvicorn>=0.35.0
websockets>=11.0.0
python-dotenv>=1.0.0
langchain-core>=0.3.67
langchain-openai>=0.3.27
langchain-google-genai>=2.1.6
langgraph>=0.5.1
langsmith>=0.4.4
a2a-sdk>=0.2.10
aiohttp>=3.12.13
pyyaml>=6.0.1
requests>=2.32.4
